New package structure. lib/ to contain non api client code
