+++ notion_client/lib/custom_enums.py
@@ -1,0 +1,97 @@
+from enum import Enum
+
+class RichTextType(Enum):
+    text = "text"
+    mention = "mention"
+    equation = "equation"
+
+class ParentType(Enum):
+    page = "page_id"
+    database = "database_id"
+    workspace = "workspace"
+
+class BasicColor(Enum):
+    default = "default"
+    gray = "gray"
+    brown = "brown"
+    orange = "orange"
+    yellow = "yellow"
+    green = "green"
+    blue = "blue"
+    purple = "purple"
+    pink = "pink"
+    red = "red"
+
+
+class Color(Enum):
+    default = "default"
+    gray = "gray"
+    brown = "brown"
+    orange = "orange"
+    yellow = "yellow"
+    green = "green"
+    blue = "blue"
+    purple = "purple"
+    pink = "pink"
+    red = "red"
+    gray_background = "gray_background"
+    brown_background = "brown_background"
+    orange_background = "orange_background"
+    yellow_background = "yellow_background"
+    green_background = "green_background"
+    blue_background = "blue_background"
+    purple_background = "purple_background"
+    pink_background = "pink_background"
+    red_background = "red_background"
+
+
+class RollupFunctionType(Enum):
+    count_all = "count_all"
+    count_values = "count_values"
+    count_unique_values = "count_unique_values"
+    count_empty = "count_empty"
+    count_not_empty = "count_not_empty"
+    percent_empty = "percent_empty"
+    percent_not_empty = "percent_not_empty"
+    sum = "sum"
+    average = "average"
+    median = "median"
+    min = "min"
+    max = "max"
+    range = "range"
+
+
+class DatabasePropertyType(Enum):
+    title = "title"
+    rich_text = "rich_text"
+    number = "number"
+    select = "select"
+    multi_select = "multi_select"
+    date = "date"
+    people = "people"
+    file = "file"
+    checkbox = "checkbox"
+    url = "url"
+    email = "email"
+    phone_number = "phone_number"
+    formula = "formula"
+    relation = "relation"
+    rollup = "rollup"
+    created_time = "created_time"
+    created_by = "created_by"
+    last_edited_time = "last_edited_time"
+    last_edited_by = "last_edited_by"
+
+
+class NumberPropertyFormat(Enum):
+    number = "number"
+    number_with_commas = "number_with_commas"
+    percent = "percent"
+    dollar = "dollar"
+    euro = "euro"
+    pound = "pound"
+    yen = "yen"
+    ruble = "ruble"
+    rupee = "rupee"
+    won = "won"
+    yuan = "yuan"
